DATA-PROJECTS/DATA-PROJECT-3/terraform/modules/appointment_api/requirements.txt
Flask==3.0.2
psycopg2-binary>=2.9.8
python-dateutil>=2.8.1
python-dotenv>=1.0.1
APScheduler==3.10.4
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
Faker>=18.0.0
requests>=2.28.0